backend/requirements.txt
numpy==1.26.4
streamlit==1.38.0
tensorflow==2.17.0
matplotlib==3.8.4
flask==3.0.3
pandas==2.2.2
scikit-learn==1.5.0
seaborn==0.13.2
scipy==1.13.1
pillow==9.3.0
gunicorn==20.1.0  
opencv-python

